# PERF: setting PeriodDtype column