```
from asv_bench.benchmarks.period import *
self = DataFramePeriodColumn()
self.setup()

%timeit self.time_setitem_period_column()
59.8 µs ± 2.19 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master
33 µs ± 933 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```